+++ R/mp32wav.R
@@ -106,8 +106,8 @@ mp32wav <- function(samp.rate = NULL, parallel = 1, path = NULL,
   if (length(files) == 0) stop("no 'mp3' files in working directory")
   
   #exclude the ones that already have a .wav version
-  if (!overwrite) 
-    {wavs <- list.files(pattern = "\\.wav$", ignore.case = TRUE)
+  if (!overwrite) {
+    wavs <- list.files(pattern = "\\.wav$", ignore.case = TRUE)
   files <- files[!substr(files, 0, nchar(files) - 4) %in% substr(wavs, 0, nchar(wavs) - 4)]
   if (length(files) == 0) stop("all 'mp3' files have been converted")
   }